+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/CircuitStatus.java
@@ -1,0 +1,7 @@
+package com.mageddo.dnsproxyserver.solver.remote;
+
+public enum CircuitStatus {
+  OPEN,
+  CLOSED,
+  HALF_OPEN
+}
