--- src/main/java/com/mageddo/dnsproxyserver/solver/remote/application/CircuitStatus.java
@@ -1,7 +1,0 @@
-package com.mageddo.dnsproxyserver.solver.remote.application;
-
-public enum CircuitStatus {
-  OPEN,
-  CLOSED,
-  HALF_OPEN
-}
