+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/configurator/SolverRemoteModule.java
@@ -1,7 +1,7 @@
 package com.mageddo.dnsproxyserver.solver.remote.configurator;
 
-import com.mageddo.dnsproxyserver.solver.remote.CircuitBreakerService;
-import com.mageddo.dnsproxyserver.solver.remote.application.CircuitBreakerFailSafeService;
+import com.mageddo.dnsproxyserver.solver.remote.application.CircuitBreakerService;
+import com.mageddo.dnsproxyserver.solver.remote.application.failsafe.CircuitBreakerFailSafeService;
 import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAO;
 import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAOImpl;
 import dagger.Binds;
